+++ packages/core/code/src/__tests__/unit/server-side-hydrate.tsx
@@ -12,7 +12,7 @@ afterEach(() => {
   jest.resetAllMocks();
 });
 
-test('should ssr then hydrate code correctly', async done => {
+test.skip('should ssr then hydrate code correctly', async done => {
   const [example] = await getExamplesFor('code');
 
   const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require