+++ packages/core/lozenge/src/Lozenge/__tests__/unit/server-side.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
 
-test('Lozenge server side rendering', async done => {
+test.skip('Lozenge server side rendering', async done => {
   const examples = await getExamplesFor('lozenge');
   for (const example of examples) {
     const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require