+++ packages/core/width-detector/src/__tests__/unit/server-side.tsx
@@ -3,7 +3,7 @@ import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
 import * as ReactDOMServer from 'react-dom/server';
 import WidthDetector from '../..';
 
-test('Width detector server side rendering', async done => {
+test.skip('Width detector server side rendering', async done => {
   const examples = await getExamplesFor('width-detector');
   for (const example of examples) {
     const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
@@ -12,7 +12,7 @@ test('Width detector server side rendering', async done => {
   done();
 });
 
-test('Width detector should render children immediately for SSR', async () => {
+test.skip('Width detector should render children immediately for SSR', async () => {
   const markup = <div id="foo123">Foo</div>;
   const markupString = ReactDOMServer.renderToStaticMarkup(markup);
   const html = ReactDOMServer.renderToString(