+++ packages/elements/share/src/components/ShareDialogWithTrigger.tsx
@@ -30,6 +30,7 @@ import {
   copyLinkButtonClicked,
   screenEvent,
   formShareSubmitted,
+  CHANNEL_ID,
 } from './analytics';
 import ShareButton from './ShareButton';
 import { ShareForm } from './ShareForm';
@@ -49,7 +50,9 @@ export type Props = {
   config?: ConfigResponse;
   children?: RenderCustomTriggerButton;
   copyLink: string;
-  isCopyLinkShortened: boolean;
+  analyticsDecorator?: (
+    payload: AnalyticsEventPayload,
+  ) => AnalyticsEventPayload;
   dialogPlacement?: DialogPlacement;
   isDisabled?: boolean;
   isFetchingConfig?: boolean;
@@ -90,7 +93,6 @@ export class ShareDialogWithTriggerInternal extends React.Component<
   State
 > {
   static defaultProps: Partial<Props> = {
-    isCopyLinkShortened: false,
     isDisabled: false,
     dialogPlacement: 'bottom-end',
     shouldCloseOnEscapePress: true,
@@ -117,10 +119,9 @@ export class ShareDialogWithTriggerInternal extends React.Component<
   };
 
   private createAndFireEvent = (payload: AnalyticsEventPayload) => {
-    const { createAnalyticsEvent } = this.props;
-    if (createAnalyticsEvent) {
-      createAnalyticsEvent(payload).fire('fabric-elements');
-    }
+    const { createAnalyticsEvent, analyticsDecorator } = this.props;
+    if (analyticsDecorator) payload = analyticsDecorator(payload);
+    if (createAnalyticsEvent) createAnalyticsEvent(payload).fire(CHANNEL_ID);
   };
 
   private getFlags = (
@@ -266,10 +267,8 @@ export class ShareDialogWithTriggerInternal extends React.Component<
   };
 
   handleCopyLink = () => {
-    const { copyLinkOrigin, isCopyLinkShortened } = this.props;
-    this.createAndFireEvent(
-      copyLinkButtonClicked(this.start, isCopyLinkShortened, copyLinkOrigin),
-    );
+    const { copyLinkOrigin } = this.props;
+    this.createAndFireEvent(copyLinkButtonClicked(this.start, copyLinkOrigin));
   };
 
   render() {