+++ packages/elements/user-picker/src/__tests__/unit/components/BaseUserPickerSpec.tsx
@@ -98,7 +98,7 @@ describe('BaseUserPicker', () => {
     const select = component.find(Select);
     select.simulate('change', userOptions[0], { action: 'select-option' });
 
-    expect(onSelection).toHaveBeenCalledWith(options[0]);
+    expect(onSelection).toHaveBeenCalledWith(options[0], undefined);
   });
 
   it('should trigger props.onClear if onChange with clear action', () => {
@@ -448,6 +448,64 @@ describe('BaseUserPicker', () => {
           expect(callback).toHaveBeenCalledWith(...payload);
         },
       );
+
+      it('should pass session id on select when it starts opened', () => {
+        const onSelection = jest.fn();
+        const component = mountWithIntl(
+          getBasePicker({ onSelection, open: true }),
+        );
+        const input = component.find(Select);
+        input
+          .props()
+          ['onChange']({ data: 'user-id' }, { action: 'select-option' });
+        expect(onSelection).toHaveBeenCalledWith(
+          'user-id',
+          'random-session-id',
+        );
+      });
+
+      it('should pass session id on focus before open', () => {
+        const onFocus = jest.fn();
+        const component = mountWithIntl(getBasePicker({ onFocus }));
+        const input = component.find('input');
+        input.simulate('focus');
+        expect(onFocus).toHaveBeenCalledWith('random-session-id');
+      });
+
+      it('should use the same session id on 2nd focus', async () => {
+        analyticsSpy
+          .mockReturnValueOnce({ id: 'session-first' })
+          .mockReturnValueOnce({ id: 'session-second' });
+        const onFocus = jest.fn();
+        const component = mountWithIntl(getBasePicker({ onFocus }));
+        const input = component.find('input');
+        input.simulate('focus');
+        await component.update();
+        input.simulate('focus');
+        await component.update();
+        expect(onFocus).toHaveBeenCalledTimes(2);
+        expect(onFocus).toHaveBeenCalledWith('session-first');
+      });
+
+      it('should use new session id for on focus if open is false', async () => {
+        analyticsSpy
+          .mockReturnValueOnce({ id: 'session-first' })
+          .mockReturnValueOnce({ id: 'session-second' });
+        const onFocus = jest.fn();
+        const component = mountWithIntl(
+          getBasePicker({ onFocus, open: false }),
+        );
+        const input = component.find('input');
+        input.simulate('focus');
+        await component.update();
+        input.simulate('focus');
+        await component.update();
+        expect(onFocus).toHaveBeenCalledTimes(2);
+        expect(onFocus.mock.calls).toMatchObject([
+          ['session-first'],
+          ['session-second'],
+        ]);
+      });
     });
   });