+++ packages/navigation/atlassian-switcher/examples/21-confluence-should-not-display-xsell.tsx
@@ -57,11 +57,7 @@ class ConfluenceSwitcherExample extends React.Component {
     return (
       <div style={{ padding: '2rem' }}>
         <Drawer onClose={this.onClose} isOpen={this.state.isDrawerOpen}>
-          <AtlassianSwitcher
-            product="confluence"
-            cloudId="some-cloud-id"
-            triggerXFlow={this.onTriggerXFlow}
-          />
+          <AtlassianSwitcher product="confluence" cloudId="some-cloud-id" />
         </Drawer>
         <Button type="button" onClick={this.openDrawer}>
           Open drawer