+++ packages/core/feedback-collector/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @atlaskit/feedback-collector
 
+## 4.0.8
+
+- Updated dependencies [87a2638655](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/87a2638655):
+  - @atlaskit/button@13.0.10
+  - @atlaskit/form@6.1.2
+  - @atlaskit/modal-dialog@10.0.8
+  - @atlaskit/select@9.1.10
+  - @atlaskit/checkbox@9.0.0
+
+## 4.0.7
+
+- Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):
+  - @atlaskit/docs@8.1.3
+  - @atlaskit/button@13.0.9
+  - @atlaskit/checkbox@8.0.5
+  - @atlaskit/form@6.1.1
+  - @atlaskit/flag@12.0.10
+  - @atlaskit/global-navigation@7.3.1
+  - @atlaskit/modal-dialog@10.0.7
+  - @atlaskit/navigation-next@6.3.2
+  - @atlaskit/select@9.1.8
+  - @atlaskit/icon@19.0.0
+
 ## 4.0.6
 
 ### Patch Changes