+++ packages/core/layer-manager/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/layer-manager
 
+## 7.0.8
+
+- Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):
+  - @atlaskit/docs@8.1.3
+  - @atlaskit/button@13.0.9
+  - @atlaskit/modal-dialog@10.0.7
+  - @atlaskit/onboarding@8.0.6
+  - @atlaskit/section-message@4.0.5
+  - @atlaskit/tooltip@15.0.2
+  - @atlaskit/icon@19.0.0
+
 ## 7.0.7
 
 ### Patch Changes