+++ packages/core/table-tree/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atlaskit/table-tree
 
+## 7.0.6
+
+- Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):
+  - @atlaskit/docs@8.1.3
+  - @atlaskit/button@13.0.9
+  - @atlaskit/section-message@4.0.5
+  - @atlaskit/select@9.1.8
+  - @atlaskit/icon@19.0.0
+
 ## 7.0.5
 
 ### Patch Changes