+++ packages/search/quick-search/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/quick-search
 
+## 7.5.1
+
+- Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):
+  - @atlaskit/docs@8.1.3
+  - @atlaskit/avatar@16.0.6
+  - @atlaskit/drawer@4.2.1
+  - @atlaskit/field-base@13.0.6
+  - @atlaskit/item@10.0.5
+  - @atlaskit/navigation@35.1.8
+  - @atlaskit/icon@19.0.0
+
 ## 7.5.0
 
 ### Minor Changes